requirements.txt
flask
openai
flask-cors
requests
python-dotenv
pyodbc
httpx
mcp[cli]